+++ scripts/monorepo/__tests__/find-and-publish-all-bumped-packages-test.js
@@ -9,12 +9,12 @@
 
 const {spawnSync} = require('child_process');
 
-const {BUMP_COMMIT_MESSAGE} = require('../monorepo/constants');
-const forEachPackage = require('../monorepo/for-each-package');
-const findAndPublishAllBumpedPackages = require('../monorepo/find-and-publish-all-bumped-packages');
+const {BUMP_COMMIT_MESSAGE} = require('../constants');
+const forEachPackage = require('../for-each-package');
+const findAndPublishAllBumpedPackages = require('../find-and-publish-all-bumped-packages');
 
 jest.mock('child_process', () => ({spawnSync: jest.fn()}));
-jest.mock('../monorepo/for-each-package', () => jest.fn());
+jest.mock('../for-each-package', () => jest.fn());
 
 describe('findAndPublishAllBumpedPackages', () => {
   it('throws an error if updated version is not 0.x.y', () => {
